requirements.txt
channels==3.0.3
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
